enterprise_ai_chat_rag_demo/requirements.txt
openai>=1.75.0
streamlit>=1.44.0
python-dotenv>=1.0.1
